+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Metadata` class providing description, unit, and other information for BayBE objects
+
 ### Fixed
 - The Python version specifier now also allows patch versions of Python 3.13